chore(garuda): decommission the media indexer — its Drive corpus was never filled
Zero's call, on a measurement: the GARUDA Drive tree has **0 files** and **0 in
the trash**, and the account's entire trash holds one unrelated file. So the
corpus was not emptied — it was never filled. Seven carefully named subfolders
(published, research, drafts, intelligence, audio, videos, photos), all seven
matching the hardcoded IDs, all empty. Qdrant `garuda_assets`: 0 points. Two
successful runs ever, out of 78.

That also retires the last open question from last night: the 24,687 tombstones
of the first working run were never "the corpus disappearing". They came from
the rest of that Drive, which is WR2's working account (~40 `WR2-<uuid>` folders
in root, created and destroyed per carousel) — exactly the noise #3708 taught
the indexer to ignore.

**Four different things in this organism are called GARUDA. Three are alive and
untouched here:** MATA GARUDA (intelligence, `com.matagaruda.*`), the OSINT-Nexus
Neo4j bridges (`garuda-consumer.sh`, `garuda-gap-detector.sh` — the only garuda
entries in declared-pairs.json), and GARUDA B1/VOA. Only the media indexer goes.

Removed: `indexer/`, `cli/garuda_{bootstrap,gc,indexer}.py`, `maintenance/`,
`alerts.py`, and the eight test files whose subject was the indexer.

Kept, because the magazine needs it: `handlers/image_handler.py`, moved out of
`indexer/handlers/` to `zantara_media/handlers/`. It never depended on the
indexer — stdlib and httpx only — and `magazine/media_resolver.py:352` uses
`extract_image` as its default describer. `security/` stays for the same reason.
The two tests that call `extract_image` directly moved to
`tests/test_image_handler.py`; mutation-verified (breaking the return value
fails them).

**The immune-enforcement step was WIDENED, not deleted.** It pinned exactly one
corpus, `test_alerts_gateway.py`, and its own comment says why: "nothing else
runs apps/zantara-media's tests — which is how the alert path stayed mute for
141 nights". `alerts.py` is GARUDA by construction (`--source garuda`,
`dedup_key = f"garuda:{condition}:{host}"`, "🚨 GARUDA CRITICAL"), so it goes
with the pipeline it alerted for — but deleting the step with it would have
re-opened that exact hole: magazine-auto-assets.yml runs three named magazine
files, so test_image_handler.py and test_dlp.py would have run nowhere. The step
now runs the whole remaining suite, and the trigger paths became globs so it
cannot go stale as files come and go. Verified the globs in a real `case`:
nested paths match, `apps/other/x.py` does not.

**Downstream consequence, handled rather than discovered later:** the daily
consumer of the one non-SYSTEM `google_drive_tokens` row was this indexer.
Nothing refreshes it now, so within STALE_REFRESH_DAYS the watchdog will call it
`stale-refresh` — correctly. Its docstring now says what that verdict means
("no consumer left, consider revoking"), instead of the two things it would
otherwise have claimed falsely. The grant row is left in place: revoking a
Google grant is Zero's call, not a side effect of a code cleanup.

Already done outside this diff: the two crontab lines on Pro (indexer 20:36
daily, gc Saturday 21:00) are disarmed and commented with the reason; backup at
`~/backups/crontab-pre-garuda-removal-20260807-115852.txt`, 256 lines before and
after, 0 active garuda-indexer/gc lines, and all 27 garuda-named LaunchAgents
(MATA GARUDA + Nexus) still loaded.

Verified: actionlint clean; 195 tests pass, both from the app dir and from the
repo root exactly as the widened gate invokes them; watchdog corpus 25/25.
