+++ .github/workflows/immune-enforcement.yml
@@ -170,8 +170,8 @@ jobs:
               scripts/drive_token_watchdog.py|\
               scripts/tests/test_drive_watchdog_fly_path.py|\
               scripts/tests/test_drive_watchdog_tier_ratchet.py|\
-              apps/zantara-media/zantara_media/alerts.py|\
-              apps/zantara-media/tests/test_alerts_gateway.py|\
+              apps/zantara-media/zantara_media/*|\
+              apps/zantara-media/tests/*|\
               scripts/tests/test_ci_service_images_use_mirror.py|\
               scripts/tests/test_lint_workflow_errexit_decap.py|\
               scripts/lint_workflow_errexit_decap.py|\
@@ -267,19 +267,26 @@ jobs:
           done
           exit $fail
 
-      - name: garuda alert gateway guilt+innocence (apps/zantara-media)
+      - name: apps/zantara-media suite (replaces the garuda alert-gateway pin)
         if: steps.paths.outputs.relevant == 'true'
         # Its own step, not the loop above: the loop runs `pytest <path>` from
         # the repo root and this corpus imports `zantara_media`, which lives
-        # under apps/. Both files ARE sentinel trigger paths, so a PR that
-        # guts either one runs this.
+        # under apps/.
         #
-        # DECLARED GAP: test_immune_enforcement_trigger_symmetry.py only reads
-        # the loop's `scripts/**.py` entries, so it cannot vouch for this step.
-        # Armed here at birth precisely because nothing else runs
-        # apps/zantara-media's tests — no workflow did, which is how the alert
-        # path stayed mute for 141 nights.
-        run: PYTHONPATH=apps/zantara-media python -m pytest apps/zantara-media/tests/test_alerts_gateway.py -q
+        # Widened 2026-08-07 with the GARUDA decommission. This step was born
+        # pinning ONE corpus (test_alerts_gateway.py) because nothing else ran
+        # apps/zantara-media's tests — "which is how the alert path stayed mute
+        # for 141 nights". That module is gone with the pipeline it alerted for,
+        # and deleting the step with it would have re-opened exactly the hole
+        # the comment describes: magazine-auto-assets.yml runs three named
+        # magazine files, so test_image_handler.py and test_dlp.py would have
+        # run nowhere. Running the whole remaining suite is strictly more than
+        # the pin it replaces, and it cannot go stale as files come and go.
+        #
+        # DECLARED GAP (unchanged): test_immune_enforcement_trigger_symmetry.py
+        # only reads the loop's `scripts/**.py` entries, so it cannot vouch for
+        # this step.
+        run: PYTHONPATH=apps/zantara-media python -m pytest apps/zantara-media/tests -q
 
       - name: PII gate merge-scope guilt+innocence (.husky/pre-commit)
         if: steps.paths.outputs.relevant == 'true'
