--- apps/zantara-media/tests/test_atomic.py
@@ -1,192 +1,0 @@
-"""Tests for atomic ordering: Qdrant upsert FIRST, then Postgres.
-
-Per spec: if Qdrant fails, Postgres must never be written.
-"""
-
-from __future__ import annotations
-
-from datetime import UTC, datetime
-from unittest.mock import AsyncMock, MagicMock
-
-import pytest
-
-from zantara_media.indexer.drive_client import DriveFile
-from zantara_media.indexer.pipeline import Pipeline
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
-def _make_file(
-    file_id: str = "file_atomic",
-    name: str = "report.pdf",
-    mime_type: str = "application/pdf",
-    size: int = 2048,
-    parents: list[str] | None = None,
-) -> DriveFile:
-    return DriveFile(
-        id=file_id,
-        name=name,
-        mime_type=mime_type,
-        parents=parents or ["1n3VjN-YZGGH-6-yByxIi0rLGxi4iTDu1"],
-        size=size,
-        modified_time=datetime(2024, 6, 15, tzinfo=UTC),
-        version=3,
-        trashed=False,
-    )
-
-
-def _make_pipeline(
-    *,
-    qdrant_side_effect: Exception | None = None,
-) -> tuple[Pipeline, MagicMock, MagicMock, MagicMock]:
-    """Return (pipeline, embedder_mock, qdrant_mock, pg_mock)."""
-    embedder = MagicMock()
-    embedder.embed_text = AsyncMock(return_value=[0.0] * 1536)
-
-    qdrant = MagicMock()
-    if qdrant_side_effect is not None:
-        qdrant.upsert = AsyncMock(side_effect=qdrant_side_effect)
-    else:
-        qdrant.upsert = AsyncMock(return_value=None)
-
-    pg = MagicMock()
-    pg.check_content_hash = AsyncMock(return_value=None)
-    pg.upsert_index_record = AsyncMock(return_value=None)
-    pg.mark_quarantined = AsyncMock(return_value=None)
-    pg.mark_archived = AsyncMock(return_value=None)
-
-    drive = MagicMock()
-    drive.download_file = AsyncMock(return_value=b"%PDF-1.4 test atomic content")
-
-    pipeline = Pipeline(
-        embedder=embedder,
-        qdrant_writer=qdrant,
-        postgres_writer=pg,
-        drive_client=drive,
-    )
-    return pipeline, embedder, qdrant, pg
-
-
-# ---------------------------------------------------------------------------
-# Test 1: Happy path — Qdrant OK → Postgres called
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_postgres_called_after_qdrant_success() -> None:
-    """When Qdrant upsert succeeds, Postgres upsert_index_record is called."""
-    pipeline, embedder, qdrant, pg = _make_pipeline()
-    file = _make_file()
-
-    result = await pipeline.index_file_safe(file)
-
-    assert result.status == "indexed"
-    qdrant.upsert.assert_called_once()
-    pg.upsert_index_record.assert_called_once()
-
-    # Verify call ordering: Qdrant must be called before Postgres.
-    # We confirm by checking the manager mock_calls ordering on the
-    # individual mocks (both called exactly once, and qdrant was first
-    # since pg would not be reached if qdrant raised).
-    qdrant_call_count = qdrant.upsert.call_count
-    pg_call_count = pg.upsert_index_record.call_count
-    assert qdrant_call_count == 1
-    assert pg_call_count == 1
-
-
-# ---------------------------------------------------------------------------
-# Test 2: Qdrant fails → Postgres NOT called
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_postgres_not_called_when_qdrant_fails() -> None:
-    """When Qdrant raises an exception, Postgres upsert_index_record is NEVER called."""
-    pipeline, embedder, qdrant, pg = _make_pipeline(
-        qdrant_side_effect=RuntimeError("Qdrant connection timeout")
-    )
-    file = _make_file()
-
-    result = await pipeline.index_file_safe(file)
-
-    # Pipeline must catch the error and return status="error" (no re-raise)
-    assert result.status == "error"
-    assert result.reason is not None
-
-    # CRITICAL: Postgres must not have been written
-    pg.upsert_index_record.assert_not_called()
-
-
-# ---------------------------------------------------------------------------
-# Test 3: PipelineResult contains error status when Qdrant fails
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_pipeline_result_error_on_qdrant_failure() -> None:
-    """PipelineResult.status == 'error' when Qdrant raises."""
-    pipeline, _, qdrant, pg = _make_pipeline(
-        qdrant_side_effect=ConnectionError("network unreachable")
-    )
-    file = _make_file(file_id="file_error_test")
-
-    result = await pipeline.index_file_safe(file)
-
-    assert result.status == "error"
-    assert result.file_id == "file_error_test"
-    assert result.reason is not None and len(result.reason) > 0
-    pg.upsert_index_record.assert_not_called()
-
-
-# ---------------------------------------------------------------------------
-# Test 4: Error isolation — other files in the batch are unaffected
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_error_isolation_in_batch() -> None:
-    """A failure in one file does not prevent other files from being indexed."""
-    call_count = 0
-
-    async def qdrant_side_effect_alternating(*args, **kwargs):  # type: ignore[no-untyped-def]
-        nonlocal call_count
-        call_count += 1
-        if call_count == 1:
-            raise RuntimeError("first file Qdrant error")
-        # Second call succeeds
-
-    embedder = MagicMock()
-    embedder.embed_text = AsyncMock(return_value=[0.1] * 1536)
-
-    qdrant = MagicMock()
-    qdrant.upsert = AsyncMock(side_effect=qdrant_side_effect_alternating)
-
-    pg = MagicMock()
-    pg.check_content_hash = AsyncMock(return_value=None)
-    pg.upsert_index_record = AsyncMock(return_value=None)
-    pg.mark_quarantined = AsyncMock(return_value=None)
-
-    drive = MagicMock()
-    drive.download_file = AsyncMock(return_value=b"%PDF-1.4 batch isolation test")
-
-    pipeline = Pipeline(
-        embedder=embedder,
-        qdrant_writer=qdrant,
-        postgres_writer=pg,
-        drive_client=drive,
-    )
-
-    file_1 = _make_file(file_id="file_fail", name="fail.pdf")
-    file_2 = _make_file(file_id="file_ok", name="ok.pdf")
-
-    result_1 = await pipeline.index_file_safe(file_1)
-    result_2 = await pipeline.index_file_safe(file_2)
-
-    # First file failed (Qdrant error)
-    assert result_1.status == "error"
-
-    # Second file succeeded — error isolation confirmed
-    assert result_2.status == "indexed"
-    assert pg.upsert_index_record.call_count == 1  # only for file_2
