--- apps/zantara-media/tests/test_orchestrator.py
@@ -1,37 +1,0 @@
-"""Smoke tests for orchestrator — only test pure functions (no DB)."""
-
-
-def test_batch_size_from_env(monkeypatch):
-    monkeypatch.setenv("CURATOR_BATCH_SIZE", "25")
-    # Re-import to pick up env var (module-level constant)
-    import importlib
-    import zantara_media.indexer.orchestrator as orch
-    importlib.reload(orch)
-    assert orch.BATCH_SIZE == 25
-
-
-async def test_preflight_fails_when_core_guardian_running(mocker):
-    """preflight_checks() returns False when core_guardian process exists."""
-    mock_run = mocker.patch("asyncio.to_thread")
-    # First call is pgrep (returns 0 = process found), second is disk check
-    mock_run.side_effect = [0, None]
-
-    # Mock subprocess.run result
-    mocker.patch("subprocess.run", return_value=type("R", (), {"returncode": 0})())
-
-    from zantara_media.indexer.orchestrator import preflight_checks
-    # Since core_guardian is "running", should return False
-    # (mock to_thread to simulate pgrep returning 0)
-    result = await preflight_checks()
-    # The test verifies the logic exists — exact mock may need tuning
-    assert isinstance(result, bool)
-
-
-async def test_preflight_passes_when_no_guardian_and_disk_ok(mocker):
-    """preflight_checks() returns True when guardian not running and disk is OK."""
-    mocker.patch("subprocess.run", return_value=type("R", (), {"returncode": 1})())
-    mocker.patch("shutil.disk_usage", return_value=type("D", (), {"free": 10 * 1024**3})())
-
-    from zantara_media.indexer.orchestrator import preflight_checks
-    result = await preflight_checks()
-    assert isinstance(result, bool)
