+++ apps/zantara-media/zantara_media/handlers/image_handler.py
