--- apps/zantara-media/zantara_media/indexer/__init__.py
@@ -1,1 +1,0 @@
-"""GARUDA Indexer — Drive change-feed consumer."""
